backend/pyproject.toml
[project]
name = "fixfirst-edge-backend"
version = "0.1.0"
description = "Offline multimodal maintenance copilot backend"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "pydantic==2.9.2",
    "python-multipart==0.0.12",
    "pdfplumber==0.11.4",
    "sentence-transformers==3.2.1",
    "faster-whisper==1.0.3",
    "numpy==2.1.2",
    "pandas==2.2.3",
    "python-dotenv==1.0.1",
    "pytest==8.3.3",
    "httpx==0.27.2",
    "actian-vectorai==0.1.0b2",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
